+++ src/users/delete_user.php
@@ -1,31 +1,32 @@
 <?php
-session_start();
-// if (!isset($_SESSION['username'])) {
-//     header("Location: login.html");
-//     exit();
-// }
+    session_start();
 
-// Database configuration and connection 
-include "../database/database_connection.php"; // Defines the variables :  $servername, $username, $password, $dbname
+    if (!isset($_SESSION['id'])) {
+        header("Location: login.php");
+        exit();
+    }
 
-// Get user ID from URL
-$id = intval($_GET['id']);
+    // Database configuration and connection
+    include "../database/database_connection.php"; // Defines the variables :  $servername, $username, $password, $dbname
 
-// Delete user
-$sql = "DELETE FROM Students WHERE id='$id'";
+    // Get user ID from URL
+    $id = intval($_GET['id']);
 
-if ($conn->query($sql) === TRUE) 
-{
-    // Redirect using JavaScript to go back 1 page
-    echo '<script type="text/javascript"> window.history.go(-1); </script>';
-    exit();
-} 
-else 
-{
-    echo '<script>alert("Can\'t delete this student, because he has borrowed books !");</script>';
-    echo '<script language="JavaScript" type="text/javascript">history.go(-1);</script>';
-    // echo "<p style='text-align: center;'>Can't delete this student, because he has borrowed books !</p>";
-}
+    // Delete user
+    $sql = "DELETE FROM Students WHERE id='$id'";
 
-$conn->close();
+    if ($conn->query($sql) === TRUE) 
+    {
+        // Redirect using JavaScript to go back 1 page
+        echo '<script type="text/javascript"> window.history.go(-1); </script>';
+        exit();
+    } 
+    else 
+    {
+        echo '<script>alert("Can\'t delete this student, because he has borrowed books or he had borrowed a book and his ID is in the Database. !");</script>';
+        echo '<script language="JavaScript" type="text/javascript">history.go(-1);</script>';
+        // echo "<p style='text-align: center;'>Can't delete this student, because he has borrowed books !</p>";
+    }
+
+    $conn->close();
 ?>
